==> processing/02-modelo-lcmm.Rmd <==

```{r, include=FALSE}

# Librerías básicas
library(tidyverse)
library(lcmm)

```


# Trayectorias de sintomatologia depresiva. Modelo Latent Class Mixed Model (LCMM)

## Identificar trayectorias:

### Modelos de trayectorias lineales: {-}

Se usa el modelo de 1 clase para fijar los valores iniciales a iterar:

```{r modelos-lcmm-lin, results=FALSE, cache = TRUE}

load(file.path('..', 'inputs', 'modelos_lcmm.RData'))

```

<!-- ### Modelos de trayectorias cuadraticas: {-} -->

<!-- ```{r modelos-lcmm-quad, results=FALSE, cache=TRUE} -->

<!-- lcmm1_quad <- hlme(phq9 ~ ola + I(ola^2), subject = "idencuesta", ng = 1, -->
<!--               data = elsoc_salud_long) -->

<!-- lcmm2_quad <- gridsearch(rep = 10, maxiter = 10, minit = lcmm1_quad, -->
<!--                     hlme(phq9 ~ ola +  I(ola^2), mixture = ~ ola +  I(ola^2), subject = "idencuesta", ng = 2, -->
<!--                          data = elsoc_salud_long)) -->

<!-- lcmm3_quad <- gridsearch(rep = 10, maxiter = 10, minit = lcmm1_quad, -->
<!--                     hlme(phq9 ~ ola +  I(ola^2), mixture = ~ ola +  I(ola^2), subject = "idencuesta", ng = 3, -->
<!--                          data = elsoc_salud_long)) -->

<!-- lcmm4_quad <- gridsearch(rep = 10, maxiter = 10, minit = lcmm1_quad, -->
<!--                     hlme(phq9 ~ ola +  I(ola^2), mixture = ~ ola +  I(ola^2), subject = "idencuesta", ng = 4, -->
<!--                          data = elsoc_salud_long)) -->

<!-- lcmm5_quad <- gridsearch(rep = 10, maxiter = 10, minit = lcmm1_quad, -->
<!--                     hlme(phq9 ~ ola +  I(ola^2), mixture = ~ ola +  I(ola^2), subject = "idencuesta", ng = 5, -->
<!--                          data = elsoc_salud_long)) -->

<!-- lcmm6_quad <- gridsearch(rep = 10, maxiter = 10, minit = lcmm1_quad, -->
<!--                     hlme(phq9 ~ ola +  I(ola^2), mixture = ~ ola +  I(ola^2), subject = "idencuesta", ng = 6, -->
<!--                          data = elsoc_salud_long)) -->

<!-- ``` -->

### Comparación resultados {-}

```{r, results=FALSE}
# Comparar resultados:
resultados_comparados <- summarytable(lcmm3_lin, lcmm4_lin, lcmm5_lin)

```

<!-- ```{r, results=FALSE} -->
<!-- # Comparar resultados: -->
<!-- resultados_comparados <- summarytable(lcmm1_lin,  -->
<!--                              lcmm2_lin, -->
<!--                              lcmm3_lin, -->
<!--                              lcmm4_lin, -->
<!--                              lcmm5_lin, -->
<!--                              lcmm6_lin) -->

<!-- ``` -->

```{r post-prob}
postprob(lcmm3_lin)

postprob(lcmm4_lin)

postprob(lcmm5_lin)


```


```{r}
kableExtra::kbl(resultados_comparados[,1:4],
                digits = 1)

kableExtra::kbl(resultados_comparados[,5:9],
                digits = 1)

```

## Principales trayectorias

Según criterios de información debería estar entre 4 y 5 clases:


```{r, cache = TRUE}
summary(lcmm4_lin)

pred_lcmm4_lin <- predictY(lcmm4_lin, data.frame(ola = 1:4))

plot(pred_lcmm4_lin, ylim = c(0, 27))

```

<!-- ```{r, cache = TRUE} -->
<!-- summary(lcmm4_quad) -->

<!-- pred_lcmm4_quad <- predictY(lcmm4_quad, data.frame(ola = 1:4)) -->

<!-- plot(pred_lcmm4_quad, ylim = c(0, 27)) -->

<!-- ``` -->


```{r, cache = TRUE}
summary(lcmm5_lin)

pred_lcmm5_lin <- predictY(lcmm5_lin, data.frame(ola = 1:4))

plot(pred_lcmm5_lin, ylim = c(0, 27))

```


<!-- ```{r, cache = TRUE} -->
<!-- summary(lcmm5_quad) -->

<!-- pred_lcmm5_quad <- predictY(lcmm5_quad, data.frame(ola = 1:4)) -->

<!-- plot(pred_lcmm5_quad, ylim = c(0, 27)) -->

<!-- ``` -->
