==> processing/04-modelo-multinomial.Rmd <==


```{r}
library(tidyverse)

# Librerias graficos
library(ggpubr)
library(plotly)
library(ggstance)

# Libreria analisis estadístico
library(nnet)
library(modelsummary)
library(marginaleffects)

```

# Modelo multinomial logit

```{r}
load(file.path('..', 'inputs', 'datos_elsoc_preparados.RData'))
load(file.path('..', 'inputs', 'datos_salud_modelo.RData'))
load(file.path('..', 'inputs', 'modelos_lcmm.RData'))

elsoc_salud_modelo <- elsoc_salud_modelo %>% 
  left_join(elsoc_lcmm, by = 'idencuesta')

```

## Comparación trayectorias

```{r}

with(elsoc_salud_modelo, rcompanion::cramerV(trayectoria_gmm.4lin, trayectoria_gmm.5lin, bias.correct = TRUE))


```

  
### Trayectorias según clasificación {-}

```{r, cache = TRUE}


elsoc_salud_long %>% 
  left_join(elsoc_lcmm, by = 'idencuesta') %>% 
  filter(!is.na(trayectoria_gmm.4lin)) %>% 
  group_by(trayectoria_gmm.4lin) %>% 
  sjlabelled::as_label(ola) %>% 
  ggplot(aes(y = phq9, x = ola, group = idencuesta, color = trayectoria_gmm.4lin)) + 
  facet_wrap(~trayectoria_gmm.4lin) +
  geom_line(alpha = .25) + 
  geom_point(alpha = .25) + 
  geom_hline(yintercept = 10, linetype  = 2) +
  scale_y_continuous(limits = c(0, 27)) +
  theme(legend.position = 'none') + 
  ylab('Indice PHQ-9') + 
  xlab(NULL)

```

## Variable dependiente: Trayectorias LCMM 4 clases - sin trayectoria salud subjetiva

```{r}

m1 <- elsoc_salud_modelo %>%
  sjlabelled::as_label(m0_sexo_w01, proyecto_vida_w03) %>% 
  nnet::multinom(trayectoria_gmm.4lin ~ m0_sexo_w01 + m0_edad_w01 + log_ing + trayectoria_deuda + trayectoria_apoyo_social + trayectoria_ocup,
           data = ., weights = ponderador02_w01, quiet = TRUE) 

m2 <- elsoc_salud_modelo %>%
  sjlabelled::as_label(m0_sexo_w01, proyecto_vida_w03) %>% 
  nnet::multinom(trayectoria_gmm.4lin ~ m0_sexo_w01 + m0_edad_w01 + log_ing + trayectoria_deuda + trayectoria_apoyo_social + trayectoria_ocup + proyecto_vida_w03 + estresor_n,
           data = ., weights = ponderador02_w01, quiet = TRUE) 

m3 <- elsoc_salud_modelo %>%
  sjlabelled::as_label(m0_sexo_w01, proyecto_vida_w03) %>% 
  nnet::multinom(trayectoria_gmm.4lin ~ m0_sexo_w01 + m0_edad_w01 + log_ing + trayectoria_deuda + trayectoria_apoyo_social + trayectoria_ocup + proyecto_vida_w03 + estresor_n + extraversion_w02 + agreeableness_w02 + conscientiousness_w02 + neuroticism_w02 + openness_w02,
           data = ., weights = ponderador02_w01, quiet = TRUE)

```

```{r}
em1 <- marginaleffects(m1, type = "probs",  conf_level = .95) %>% 
  summary() %>% 
  mutate(term = paste0(term, contrast))

em2 <- marginaleffects(m2, type = "probs",  conf_level = .95) %>% 
  summary() %>% 
  mutate(term = paste0(term, contrast))

em3 <- marginaleffects(m3, type = "probs",  conf_level = .95) %>% 
  summary() %>% 
  mutate(term = paste0(term, contrast))

```


```{r}
modelsummary::modelsummary(list(m1, m2, m3),
                           estimate = "{estimate}{stars}",
                           statistic = NULL,
                           stars = c('*' = .1, '**' = .05, '***' = .01),
                           shape = term + response ~ model,
                           exponentiate = TRUE)


```

## Resultados en formato grafico

```{r}
m3.coefs <- broom::tidy(m3, conf.int=TRUE)
```

### Variables de control {-}

```{r}

em3 %>% 
  filter(term %in% c('m0_sexo_w01Mujer - Hombre', 'm0_edad_w01dY/dX', 'log_ingdY/dX', 'proyecto_vida_w03Decidido direccion de vida - No decidido direccion de vida', 'estresor_ndY/dX')) %>% 
  mutate(term = factor(term, 
                       levels = c('m0_sexo_w01Mujer - Hombre', 'm0_edad_w01dY/dX', 'log_ingdY/dX', 'proyecto_vida_w03Decidido direccion de vida - No decidido direccion de vida', 'estresor_ndY/dX'),
                       labels = c('Sexo (mujer=1)', 'Edad', 'Log(ing)', 
                                  'Decidido direccion\nde vida', 'N estresores')),
         group = gsub('Trayectoria sintomas', 'Trayectoria sintomas\n', group)) %>% 

  ggplot(aes(x = estimate, y = fct_rev(term), colour = fct_rev(group))) +
  geom_pointrangeh(aes(xmin = conf.low, xmax = conf.high),
                   position = position_dodgev(height = 0.75)) + 
  geom_vline(xintercept = 0, linetype = 2, alpha = .5) +
  scale_color_discrete(guide = guide_legend(reverse = TRUE)) +
  theme(legend.position = 'top') +
  theme(legend.title = element_blank()) +
  xlab('Probabilidad predicha') + 
  ylab(NULL)

```



### Trayectoria sobrecarga deuda {-}

```{r}

em3 %>% 
  filter(grepl('trayectoria_deuda', term)) %>% 
  mutate(term = gsub("trayectoria_deuda", "", term),
         term = gsub(" - Nada/No muy sobrecargados deuda", "", term),
         group = gsub('Trayectoria sintomas', 'Trayectoria sintomas\n', group)) %>% 
  ggplot(aes(x = estimate, y = fct_rev(term), colour = fct_rev(group))) +
  geom_pointrangeh(aes(xmin = conf.low, xmax = conf.high),
                   position = position_dodgev(height = 0.75)) + 
  geom_vline(xintercept = 0, linetype = 2, alpha = .5) +
  scale_color_discrete(guide = guide_legend(reverse = TRUE)) +
  theme(legend.position = 'top') +
  theme(legend.title = element_blank()) +
  xlab('Probabilidad predicha cr a Nada/No muy sobrecargados deuda') + 
  ylab(NULL)

```

### Apoyo social {-}

```{r}

em3 %>% 
  filter(grepl('apoyo_social', term)) %>% 
  mutate(term = gsub("trayectoria_apoyo_social", "", term),
         term = gsub(" - Apoyo social estable", "", term),
         group = gsub('Trayectoria sintomas', 'Trayectoria sintomas\n', group)) %>% 

  ggplot(aes(x = estimate, y = fct_rev(term), colour = fct_rev(group))) +
  geom_pointrangeh(aes(xmin = conf.low, xmax = conf.high),
                   position = position_dodgev(height = 0.75)) + 
  geom_vline(xintercept = 0, linetype = 2, alpha = .5) +
  scale_color_discrete(guide = guide_legend(reverse = TRUE)) +
  theme(legend.position = 'top') +
  theme(legend.title = element_blank()) +
  xlab('Probabilidad predicha cr a Apoyo social estable') + 
  ylab(NULL)


```

### Trayectoria ocupacional {-}

```{r}

em3 %>% 
  filter(grepl('ocup', term)) %>% 
  mutate(term = gsub("trayectoria_ocup", "", term),
         term = gsub(" - Ocupados estables", "", term),
         group = gsub('Trayectoria sintomas', 'Trayectoria sintomas\n', group)) %>% 

  ggplot(aes(x = estimate, y = fct_rev(term), colour = fct_rev(group))) +
  geom_pointrangeh(aes(xmin = conf.low, xmax = conf.high),
                   position = position_dodgev(height = 0.75)) + 
  geom_vline(xintercept = 0, linetype = 2, alpha = .5) +
  scale_color_discrete(guide = guide_legend(reverse = TRUE)) +
  theme(legend.position = 'top') +
  theme(legend.title = element_blank()) +
  xlab('Probabilidad predicha cr a Ocupados estables') + 
  ylab(NULL)


```

### Variables de personalidad {-}

```{r}

em3 %>% 
  filter(term %in% c('extraversion_w02dY/dX', 'agreeableness_w02dY/dX', 'conscientiousness_w02dY/dX', 'neuroticism_w02dY/dX', 'openness_w02dY/dX')) %>% 
  mutate(term = gsub("_w02", "", term),
         term = str_to_title(gsub("dY/dX", "", term)),
         group = gsub('Trayectoria sintomas', 'Trayectoria sintomas\n', group)) %>% 

  ggplot(aes(x = estimate, y = fct_rev(term), colour = fct_rev(group))) +
  geom_pointrangeh(aes(xmin = conf.low, xmax = conf.high),
                   position = position_dodgev(height = 0.75)) + 
  geom_vline(xintercept = 0, linetype = 2, alpha = .5) +
  scale_color_discrete(guide = guide_legend(reverse = TRUE)) +
  theme(legend.position = 'top') +
  theme(legend.title = element_blank()) +
  xlab('Probabilidad predicha') + 
  ylab(NULL)


```
